Refactor - A well needed index.js cleanup
